Fixed bug where clock modulation check was logically inverted
I made Jailhouse do precisely the opposite of what I wanted. It was
exiting *only when* clock modulation was enabled...

Added some print statements for the future, since the original EIO
error was completely unhelpful and it took me a while to trace the
code to figure out that ioctl() was simply returning EIO from
Jailhouse.