Adjust the inmate and root configs
The Jailhouse communication region apparently does not take up any
physical address space - it is given pages from a page pool elsewhere.
So don't make room for an extra 0x1000 in the inmate config.

Also, the root config's inmate RAM region was too large, so it was
overlapping the IVSHMEM region. So I reduced it by 1 MB, and then
made sure the IVSHMEM region did not account for the extra 0x1000 and
that the size was a full MB.