mgh_demo.py now supports input files
Set up an arg parser.
Create a parameter to determine if input is a file or string.
Create a parameter for 'demo mode.'
Prevented the contents of the string/file from printing if larger than
256 bytes.
Properly close files at end of program.
Fix bug where rhash_output was not being printed when inmate output
doesn't match.