Get file input working better
Read input files as binary instead of utf-8 in case non-utf-8
characters are encountered (like in a binary file).
Get inmate to stop printing out the input, in case of a binary file
(the console output will be all messed up).
Create functions for writing input if from file or a string and change
some variable names.