Add option to locally buffer input in new custom heap space
Map 10 MB of pages of the new heap space to the inmate's virtual mem.
If specified, copy input data into a local buffer in the new heap
space, work on the data, then copy it back to the output.

The idea was to copy the input data out of a shared memory region, so
that as the data is worked on, it's not touching a shared page that
could introduce contention. But I don't think that's a big issue. The
delays I saw were due to frequency scaling coupling between the root's
cores and the inmate's core. This local buffering simply doubled the
time it took to process.

Move some variable definitions inside the loop.
Pull THROTTLE_MODE out as a global.
Fix up some type issues with some functions.
Improve the output time stats.