Update the Bazooka root config
Try to fix crashes when the root is terminated via jailhouse disable.
These fixes seem to make it better, but there are instances where it
still happens.

The improvement could be due to the fixes in the PCI cap lengths, or
possibly due to increasing the size of that first system RAM config.
The system-kernel-system chunk shouldn't matter, since it's all
allowed and kernel just shifts around in there somewhere each boot.
But I was just being safe.

Move comments inline to aid in visual comparisons to generated config.

command to regenerate:
jailhouse config create --console ttyS4 2019-10-26_1.config