Update root cell-inmate shmem protocol
Only have one data region for both input and output.
Only have one length field for the data.
Create a function to set the inmate's output data size.

The inmate and root will never write to the share memory at the same
time anyways, so we might as well have a bidirectional data region
to maximize the input and output data size.