Clear sync byte of shmem before each experiment
Add a new -c argument to clear the sync byte.
Restructure input arguments so -c can be used without a required
INPUT argument.

Ran into a problem where the sync byte was left in a different state
from the previous run, which caused a race condition on the first
iteration.