Rework frequency code
There was a bug where the calculated frequency was sometimes incorrect
due to multiplication overflow. The solution is to do the frequency
calculation in Linux, similar to what I did with the TSC counter and
calculating time.

Also, change the frequency output so that there is only one sample
that covers the duration of the entire workload to get an average
frequency across the workload.

Start creating frequency output files properly.