Update Bazooka root cell memory regions
The root cell was getting a FATAL: Invalid MMIO/RAM write for
accessing 0x87f7e000, which the old config thought was reserved. The
newly-generated config has the ACPI region at a different spot now
(not sure if the change was because I added some hardware or if it
changes on each boot).
At any rate, this should fix the error for now.